% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_pdf.R
\name{save_as_pdf}
\alias{save_as_pdf}
\title{Save an R graph as a PDF file}
\usage{
save_as_pdf(graph, file.name = "pdf-graph.pdf", width = 7, height = 7)
}
\arguments{
\item{graph}{R expression generating a graph}

\item{file.name}{output file name to write}

\item{width}{output file width}

\item{height}{output file height}
}
\description{
Save an R graph as a PDF file
}
