% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_image.R
\name{save_image}
\alias{save_image}
\title{Save an R graph as a high resolution image}
\usage{
save_image(graph, file.name = "graph.png", width = 7, height = 7, res = 500)
}
\arguments{
\item{graph}{expression generating a graph}

\item{file.name}{output file name to write}

\item{width}{output file width}

\item{height}{output file height}

\item{res}{the nominal resolution in in which will be recorded in the bitmap file}
}
\description{
Save an R graph as a high resolution image
}
