% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_image_from_pdf.R
\name{save_image_from_pdf}
\alias{save_image_from_pdf}
\title{Convert PDF page to an image}
\usage{
save_image_from_pdf(pdf.file, file.name = "image.png", pages = NULL)
}
\arguments{
\item{pdf.file}{PDF file to read}

\item{file.name}{output file name to write}

\item{pages}{PDF pages to read}
}
\description{
Convert PDF page to an image
}
\examples{

}
